PairedSystem
Some components (such as Transform and Physics) are inherently
connected, but may need to be accessed separately by other components.
I've added PairedSystem, a subclass of ISystem, for components like
this. It basically maintains two component lists instead of one. Ideally
I would have one System implementation that handled this sort of thing
with variadic templates, but I have no idea whether that's actually
possible, or how to do it.