SearchForEntityId
I added SearchForEntityId methods to both System and PairedSystem that
take a reference to an index and an entityID and increment the index
until the two match, returning true. It will reset the index to 0 when
it hits the end of the componentData array and return false if the total
number of increments exceeds the size of the array.