Working on Dynamic File Loading.
For the most part it works. There seems to be a bug where the Vertex
Shader (and possibly the Pixel Shader) are not being created properly,
which could have to do with file pathways.

Also, look into whether or not to have a bunch of ShaderResourceViews
stored by the CM, so that duplicates of a texture aren't being created.