ContentManager pt1
I've added the content manager code and made some changes to get it to
work with the engine. Most notably, Meshes and Materials are never dealt
with as pointers, but rather copied around as necessary. The struct is
more minimal than it was in the assignments, so this is preferable.

Note that this doesn't compile - I still ahve to grab some code from the
other branch.