Rendering works
So shaders are no longer going to be part of Assets since the executable
only needs the compiled shaders. Instead, they'll be in a Shaders folder
within the project, and they'll be copied to VertexShaders\ and
PixelShaders\ in the output directory.

Similarly, the project's working directory has been set to the output
directory rather than the project directory, and I've added a post-build
event to copy all the assets into the output directory.

I've also fixed several bugs, including an erroneous release of
ShaderResourceViews, bad referencing for vertex and index arrays during
mesh creation, and improper initialization of the camera.