Collision parallelization, d3dcompiler_47
The outer collision loop is now parallelized using parallel_for and a
mutex (for additions to m_collisions). This is an ~80% speedup.

I've added d3dcompiler_47.dll to the project folder and added a
post-build event to copy it to the output directory. This is so the
program works on sub-8.1 machines, and it's necessary because of the
shader reflection performed by SimpleShader. I haven't actually tested
this on a sub-8.1 system, so maybe it still won't work.