CollisionFunctions, bug fixes
There's now a CollisionFunctions class that contains a number of
collision methods that take the entity IDs of the two colliding objects
and dt and a method that takes two bitmasks and returns a function
pointer with the aforementioned signature.

Each bitmask represents a collection of collision types, and each
CollisionComponent now has one such bitmask. The bitmasks are created
with the global MakeCollisionMask function declared in
CollisionComponent.h and defined in CollisionSystem.cpp which takes a
vector of CollisionType enum objects.

The way collisions are now handled is that a vector of function pointers
is grabbed using the bitmasks of the colliding objects and
GetCollisionFunction. The collision functions are added to
m_collisionMap with LockVectors (new thread-safe vector) (also note that
this addition will not overwrite), and the pair of entityIDs are pushed
to those vectors.  After all objects are checked for collisions each
LockVector is iterated through, and the collision function is called on
each pair of colliding objects. m_collisionMap is cleared every frame,
obviously.

I'm not sure whether the overhead of m_collisionMap overweighs the
performance lost to instruction cache misses, so I might take it out at
some point.

Going forward, collision behavior may be added by doing the following:
-Add any necessary collision types to the CollisionType enum
-Write the collision method
-Add the necessary if-else to GetCollisionFunction (this is where you'll
specify which pair(s) of collidable types should use the new collision
function)
-Remember to pass the appropriate bitmask to any constructors you write