Made slight changes to the Content Manager.
Made a new m_textures vector that holds all the various textures for the
project. I made a new CreateTexture() method to create the textures for
m_textures and I call FindFilesInFolder() for it too.

Wasn't able to test as the parallel_for loop in the Collision code will
not compile for me.