Bounding box generation complete
The content manager will store MeshStores instead of Meshes. A MeshStore
contains a Mesh and a BoundingBox, with the BoundingBox's points
generated using the max and min of the mesh's vertices. Cosntructors
will now take a CollisionMask instead of the actual collision component,
and will copy the mask into the collision component in the MeshStore and
use that as the collision component.