Collision resolution, gravity, floor collision
Collision functions can now actually do things. Believe it or not, this
was not trivial to implement; I had forgotten that the function needed a
reference to the game object, and I ran into some nasty circular
dependencies while trying to make that happen. Long story short, files
that need the fully declared game class (meaning .cpps usually) include
it directly, and files that only need the forward declaration include
that (it's been moved into its own file). Similarly, the typedef for
CollisionFunction has been moved into its own file, and any headers that
need it include it. This is an absolutely disgusting pattern, and I love
it.

The transform update will now add a predefined value to the Y
acceleration when it loads it from each component. Tada, gravity.

The collision system will now check how far below 0 the min Y value is
for each aabb and offset both the min and max points and the entity's
transform by the same amount. It will negate the entity's Y velocity if
it is below 0 as well. We could get fancy in the future and pull a value
from a height map instead of using 0 if we wanted to have actual
terrain.

I also fixed an issue with the physics - an object having no rotational
velocity would cause an error in the rotation code. That code is now
encased in an if that checks whether the rotational velocity has a
non-zero magnitude.