Spatial hashing prep
Adding ClearVector and ClearArray in preparation for the spatial hashing
implementation. These are wrapper types that allow for selective
contiguous reuse of the underlying structure. ClearVector is for dynamic
sizing, ClearArray takes size as a template parameter. ClearArray also
allows for checking of duplicates on adding items, since I know I'll
need that functionality anyway.