Spatial hashing, collapse revision
Spatial hashing is in and working. The basic idea is that after
generating the list of aabbs, the grid (a ClearVector of ClearVectors of
collapsed components) will be resized and repositioned. Then the aabbs
will be sorted into the grid in parallel by translating their corner
points into 1D indices and pushing the aabb to the ClearVector at that
index. Checking for actual collisions is the same as before, except the
double for is no longer parallelized, and instead sits inside a
parallel_for that iterates through each grid bucket.

At the moment there are some caveats:
-cell counts are not calculated, they are hardcoded in a define at the
top of the file. In the future this will be determined by object density
-double collisions are not accounted for at all yet. This means that if
two colliding objects both cross a cell line the collision will be
checked and resolved twice. I'm thinking I can solve this by adding
every pair of objects that cross cell lines to an unordered_set before
checking for collision and only actually check them if the pair isn't
already in the set

All in all it's pretty good though, rendering is definitely the new
bottleneck. With the rendering system disabled I can get ~90 fps with
~20000 objects, which is very solid.

The collapse system has been revised. Instead of passing entityIDs and
handles to new vectors, m_collapsedComponents will now hold
CollapsedComponent<T>s that wrap a component type and hold an entityID
and a handle. This was done so I didn't have to figure out how to
propagate IDs and handles into the spatial grid for collisions, but it's
kind of gross, so I'm definitely gonna rework it some time soon.