Cell crossers pt1
I've added a new collection to CollisionSystem to deal with double
collisions across cell lines, m_cellCrossers. This is a ClearVector of
pairs of collapsed AABBs and ClearArrays. If an AABB crosses cell lines
it gets added to this collection instead of m_spatialHashGrid. After
checking collisions within the hash grid as normal, this collection will
be checked. Each element will be checked against all the elements in
each of its cells and against all the other elements in the cell
crossing collection.