m_registeredCollisions
Alternate solution. Successful collisions are checked against a record
of successful collisions, and if the collision is new it gets added to
that record and a resolution function gets queued.

The collection is a ClearVector of ClearVectors of unsigned ints. The
top vector is parallel with m_componentData, and the nested vectors
represent a list of entityIDs whose colliders have collided with this
particular collider.