Simplified collision system, component data struct, debug unparallelization
Bitmasks and collision function requests are gone. Instead, each
collision component will have a collision type, and the collision
system's initialization will create a list of tuples that contain two
collision types and a collision function pointer. Each bucket in the
hash grid will now contain a vector of ClearVectors rather than just a
ClearVector - this means a bucket will have a ClearVector for each
collision type. AABBs will be sorted into buckets according to their
collision type, and when it comes time to check collisions only the
pairs of types defined in the tuple collection will be checked. The
function in the tuple will be used to resolve the collision. This will
make implementing gameplay much simpler than it would have been with the
old system.

ComponentData is now a POD struct. This is mainly so that systems can
derive from it easily, if that's ever necessary.

I've added preprocessor macros that will replace some of the
parallel_for loops with regular for loops when the program is built in
debug mode, since debugging parallel code is a bitch.