Dynamic spatial grid size, ClearVector iterator fix
The cell counts for the spatial hash grid will now be properly(ish)
generated based on object density. This has increased performance at low
object counts by an order of magnitude.

I also discovered an issue with ClearVector's end method; it was passing
back the underlying vector's end, which went well past the end of the
actual data. This was causing any objects that shared entity IDs with
the objects past the end of the data to be erroneously deleted. I've
changed end to return the begin vector plus the count.