Transform GetComponent, blend description fix
The transform system will now store its handles in a vector whose
entries correspond to entity IDs rather than a map. This provided a 5%
speedup.

Chris pointed out that I wasn't initializing the particle blend
description properly, and that's why it was failing to create the blend
state with the description local rather than a member. fixed that, and
now the description is now local to Init.