FXAA, anisotropic iltering, bloom enhancements, particle depth buffer fix
We've got FXAA. Pretty straightforward, I just had to add another render
target and the necessary shaders. It barely afects the frame rate, so
it's set up for maximum quality.

All our samplers are now anisotropic rather than trilinear. We needed an
anisotropic sampler for FXAA, so I figured why not convert them all. I
think it looks a bit better.

The bright pixel filter for bloom now has a min, a max, and an exponent.
Pixels will luminosity below the min are ignored, pixels above the max
are taken at full value, and pixels that fall between the two are scaled
according to the exponent.

The blur now has weightings as well; each pixel's contribution to the
total is scaled by its distance from the target pixel.

I've added a depth stencil state for particle rendering that prohibits
writing to the depth buffer. This means the transparent parts of the
billboards no longer obscure other pixels.