Ground plane, Timeable, increased far clipping plane
We now have a ground plane. It's treated like every other game object,
but it doesn't have a collider. Because the ground collisions are
controlled by the collision system, I needed to add a gravity bool to
the physics component to allow entities to ignore gravity, otherwise the
ground would have fallen forever due to not colliding with itself. This
wouldn't have been an issue if I had decoupled the transform and physics
systems weeks ago like I wanted to.

I've added a Timeable class that allows for starting and pausing a
chrono timer, and reading out the total accumulated time. I'm using this
to track the distribution of time spent on each system; each one derives
from Timeable and starts and stops the timer at the beginning and end of
its update method. When window text is updated all four total times are
summed, then each one is divided by the total and printed.

The camera's clipping plane has been extended by a factor of ten. I had
a real hard time figuring out why the ground plane wasn't working until
I realized it was 90% outside the view frustum.