Distance fog volume intersection
Fog is now defined as a volume existing below a specified y value, and
the distance used by the fog equation is the length of the camera/vertex
vector that falls within that volume. This localizes fog to the area
above the ground. I still need to port it to the skybox and particles.